requirements.txt
streamlit>=1.30
pytest>=7.0
google-genai>=1.0.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
google-auth>=2.29.0
python-dotenv>=1.0.0
